requirements.txt
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-community==0.2.19
langgraph==0.1.10
langsmith==0.1.147
python-dotenv>=1.0.0
openai>=1.12.0
aiohttp>=3.9.0
typing-extensions>=4.8.0
pydantic>=2.0.0,<2.5.0
pydantic-core>=2.0.0,<2.5.0
beautifulsoup4>=4.12.0
selenium>=4.30.0
matplotlib>=3.5.0
webdriver-manager>=3.5.0
sentence-transformers>=2.2.0
qdrant-client>=1.9.0
tiktoken==0.9.0 
